NVIDIA: VR: SAUCE: virt: rmeda: Fix offset calculation
Currently the code calculates the size by multiplying the
number of pages by SZ_4K. However, the number of pages is
provided in a u32 variable, and therefore it can overflow.

Fix the offset calculation by explicitly casting the number
of pages from u32 into u64 before multiplication.

+++ drivers/virt/rmeda/rmeda_guest.c
@@ -750,8 +750,8 @@ struct rmeda_guest_mapping *rmeda_guest_validate_mapping(struct rmeda_guest *pri
 	 * The range base must be aligned with its size. Use this information
 	 * to determine the offset within the range
 	 */
-	range_size = mmio_range[report_id].number_of_pages * SZ_4K;
-	start_phys += start & (range_size - 1);
+	range_size = ((u64)mmio_range[report_id].number_of_pages) * SZ_4K;
+	start_phys += (start & (range_size - 1));
 
 	mapping = kzalloc(sizeof(*mapping), GFP_KERNEL);
 	if (!mapping)
